src/main/persistence/JsonWriter.java
package persistence;

import model.StockPortfolio;
import org.json.JSONObject;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.PrintWriter;

public class JsonWriter {

    /*
    CLASS-LEVEL COMMENT: Json Writer
    Writes data to JSON file, storing the current state of the stock portfolio and its Stocks.
    */

    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // EFFECTS: constructs writer
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer , throws FileNotException if destination file can't be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes Json representation of StockPortfolio to file
    public void write(StockPortfolio stockPortfolio) {
        JSONObject json = stockPortfolio.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }

}